Update base template with new headers and logo
Significant improvements have been made on the base template, including a site title change from "Welcome" to "Snowtricks". In addition, a sticky header featuring a navigation bar with a logo and links based on user's role has been added. In the body, a container class was introduced to achieve a better page layout and styling.