Add TricksType Form
The commit introduces a new Form Type `TricksType`. This form has fields title, content, and category with the title and content as text fields and category is a special field that uses EntityType to produce a dropdown of 'Group' entities. The form type is configured to be used with the 'Tricks' entity.