Set language attribute and update navigation links
The language attribute in the HTML tag is now set to French ("fr"). In addition, the navigation links previously hard-coded with URLs are now replaced with the path helper from the Twig template engine to dynamically generate URLs. This is to ensure that the correct route is rendered, even if the routes change in the future.