Added Video entity with related migration and repository
The Video entity has been created, with relevant properties such as: id, videoHost, videoId, embedUrl and a relation with the Tricks entity. The Doctrine migration file for creating the 'video' table in the database and setting up the foreign key with 'tricks' table also has been added. A repository to handle future interaction with Video entity was also included.