Add video handling to form and event subscribers
Introduced video handling to the form and event subscribers - a new VideoFormType, new Subscriber classes for pre-set data and SumitEvent for Video. Updated the TricksType class to include videos. Also, added a videos-collection JS controller to manage the addition and removal of videos in the UI. Additionally, update the Video entity with lifecycle callbacks to set the embedUrl.