Add delete and edit functions for tricks and comments
The TricksController has been updated to include functions for deleting and editing tricks and comments in the system. This includes the validation of CSRF tokens for these actions. The MapEntity attribute has also been applied to the edit and show methods in the TricksController to make mapping between entities and routes easier and more efficient.