Refactor form submission flow in TricksController
The form submission flow in TricksController has been refactored to handle invalid form submission separately. Previously, we only accounted for when forms were submitted and valid. Now, in case a form is submitted and not valid, a new response is built with HTTP_BAD_REQUEST status. Consequently, the 'use' statement for Doctrine\ORM\Mapping\Entity which is no longer required, has been removed.